fix: resolve std::max mixed-type overload error on MSVC (pic.cpp:209)
Agent-Logs-Url: https://github.com/Crloxa/CNProject1_266/sessions/3216b08a-4295-459f-9681-05186b8d585c

+++ pic.cpp
@@ -206,7 +206,7 @@ namespace ImgParse
 			float minRatio = std::numeric_limits<float>::max();
 			for (int i = 0; i < 4; ++i)
 			{
-				const float dist = std::max(norm(candidates[i].center - exactCenter), 1.0f);
+				const float dist = static_cast<float>(std::max(norm(candidates[i].center - exactCenter), 1.0));
 				// Bottom-right marker tends to look relatively smaller under perspective and layout,
 				// so its area-to-distance^2 ratio is usually the smallest among four corners.
 				const float ratio = candidates[i].area / (dist * dist);
